airflow/jobs/requirements.txt
pyspark==3.5.0
boto3==1.28.62
mwviews==0.2.1
python-dotenv==1.0.0